Add missing property types for styled-jsx
